requirements.txt
pymupdf
chromadb
langchain
langchain-community
langchain-openai
langchain-text-splitters
sentence-transformers
gradio
openai
groq
python-dotenv
